// src/types/sales.ts
export type SalesRecord = {
  id: string
  unidade: string
  data: string
  codigo: string
  produto: string
  nv_produto_superior_setor: string
  modalidade: string
  t_consumidor: string
  vr_un: number
  qtd: number
  faturamento: number
  vr_acres: number
  vr_liquido: number
  perc_vr_vendido: number
  mtc: string
  created_at: string
}